Feat: some more refactor
